backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
trafilatura==2.0.0
newspaper3k==0.2.8
lxml==5.4.0
lxml_html_clean==0.4.3
anthropic==0.67.0
python-docx==1.1.2
pydantic==2.11.7
